+++ src/Header.jsx
@@ -1,18 +1,19 @@
 import { 
     SearchIcon,
-    API
+    API,
+    headerPadding
 } from "./Barrel.jsx";
 
 function DashboardHeader() {
     return (
-        <div className="flex flex-auto justify-between items-center w-full py-3 px-4">
-            <form action={API + 'getdata'} method="GET" className="flex items-center border border-gray-300 rounded-md px-4 py-2">
+        <div className={"flex flex-auto justify-between items-center w-full " + headerPadding}>
+            <form action={API + 'getdata'} method="GET" className="flex flex-auto max-w-sm items-center justify-between border border-gray-300 rounded-md px-4 py-2 me-10">
                 <input
                     type="text"
                     placeholder="Search Here"
                     className="active:outline-none focus:outline-none"
                 />
-                <button type="submit" className="flex items-center justify-center h-5 w-5">
+                <button type="submit" className="flex items-center h-5 w-5">
                     <SearchIcon />
                 </button>
             </form>